Moving LinkLabel ListView subclasses to their own files
