+++ src/System.Windows.Forms/src/System/Windows/Forms/LinkLabel.cs
@@ -22,7 +22,7 @@ namespace System.Windows.Forms
     [DefaultEvent(nameof(LinkClicked))]
     [ToolboxItem("System.Windows.Forms.Design.AutoSizeToolboxItem," + AssemblyRef.SystemDesign)]
     [SRDescription(nameof(SR.DescriptionLinkLabel))]
-    public class LinkLabel : Label, IButtonControl
+    public partial class LinkLabel : Label, IButtonControl
     {
         private static readonly object s_eventLinkClicked = new object();
         private static Color s_iedisabledLinkColor = Color.Empty;
@@ -2011,780 +2011,6 @@ namespace System.Windows.Forms
                 default:
                     base.WndProc(ref msg);
                     break;
-            }
-        }
-
-        public class LinkCollection : IList
-        {
-            private readonly LinkLabel _owner;
-
-            ///  A caching mechanism for key accessor
-            ///  We use an index here rather than control so that we don't have lifetime
-            ///  issues by holding on to extra references.
-            ///  Note this is not Thread Safe - but WinForms has to be run in a STA anyways.
-            private int lastAccessedIndex = -1;
-
-            public LinkCollection(LinkLabel owner)
-            {
-                _owner = owner ?? throw new ArgumentNullException(nameof(owner));
-            }
-
-            public virtual Link this[int index]
-            {
-                get
-                {
-                    return (Link)_owner._links[index];
-                }
-                set
-                {
-                    _owner._links[index] = value;
-
-                    _owner._links.Sort(s_linkComparer);
-
-                    _owner.InvalidateTextLayout();
-                    _owner.Invalidate();
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get => this[index];
-                set
-                {
-                    if (value is Link link)
-                    {
-                        this[index] = link;
-                    }
-                    else
-                    {
-                        throw new ArgumentException(SR.LinkLabelBadLink, nameof(value));
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the child control with the specified key.
-            /// </summary>
-            public virtual Link this[string key]
-            {
-                get
-                {
-                    // We do not support null and empty string as valid keys.
-                    if (string.IsNullOrEmpty(key))
-                    {
-                        return null;
-                    }
-
-                    // Search for the key in our collection
-                    int index = IndexOfKey(key);
-                    if (IsValidIndex(index))
-                    {
-                        return this[index];
-                    }
-                    else
-                    {
-                        return null;
-                    }
-                }
-            }
-
-            [Browsable(false)]
-            public int Count => _owner._links.Count;
-
-            /// <summary>
-            ///  whether we have added a non-trivial link to the collection
-            /// </summary>
-            public bool LinksAdded { get; private set; }
-
-            object ICollection.SyncRoot => this;
-
-            bool ICollection.IsSynchronized => false;
-
-            bool IList.IsFixedSize => false;
-
-            public bool IsReadOnly => false;
-
-            public Link Add(int start, int length)
-            {
-                if (length != 0)
-                {
-                    LinksAdded = true;
-                }
-
-                return Add(start, length, null);
-            }
-
-            public Link Add(int start, int length, object linkData)
-            {
-                if (length != 0)
-                {
-                    LinksAdded = true;
-                }
-
-                // check for the special case where the list is in the "magic"
-                // state of having only the default link in it. In that case
-                // we want to clear the list before adding this link.
-
-                if (_owner._links.Count == 1
-                    && this[0].Start == 0
-                    && this[0]._length == -1)
-                {
-                    _owner._links.Clear();
-                    _owner.FocusLink = null;
-                }
-
-                Link l = new Link(_owner)
-                {
-                    Start = start,
-                    Length = length,
-                    LinkData = linkData
-                };
-                Add(l);
-                return l;
-            }
-
-            public int Add(Link value)
-            {
-                if (value != null && value.Length != 0)
-                {
-                    LinksAdded = true;
-                }
-                // check for the special case where the list is in the "magic"
-                // state of having only the default link in it. In that case
-                // we want to clear the list before adding this link.
-                //
-                if (_owner._links.Count == 1
-                    && this[0].Start == 0
-                    && this[0]._length == -1)
-                {
-                    _owner._links.Clear();
-                    _owner.FocusLink = null;
-                }
-
-                // Set the owner control for this link
-                value.Owner = _owner;
-
-                _owner._links.Add(value);
-
-                if (_owner.AutoSize)
-                {
-                    LayoutTransaction.DoLayout(_owner.ParentInternal, _owner, PropertyNames.Links);
-                    _owner.AdjustSize();
-                    _owner.Invalidate();
-                }
-
-                if (_owner.Links.Count > 1)
-                {
-                    _owner._links.Sort(LinkLabel.s_linkComparer);
-                }
-
-                _owner.ValidateNoOverlappingLinks();
-                _owner.UpdateSelectability();
-                _owner.InvalidateTextLayout();
-                _owner.Invalidate();
-
-                if (_owner.Links.Count > 1)
-                {
-                    return IndexOf(value);
-                }
-                else
-                {
-                    return 0;
-                }
-            }
-
-            int IList.Add(object value)
-            {
-                if (value is Link)
-                {
-                    return Add((Link)value);
-                }
-                else
-                {
-                    throw new ArgumentException(SR.LinkLabelBadLink, nameof(value));
-                }
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                if (value is Link)
-                {
-                    Add((Link)value);
-                }
-                else
-                {
-                    throw new ArgumentException(SR.LinkLabelBadLink, nameof(value));
-                }
-            }
-
-            public bool Contains(Link link)
-            {
-                return _owner._links.Contains(link);
-            }
-
-            /// <summary>
-            ///  Returns true if the collection contains an item with the specified key, false otherwise.
-            /// </summary>
-            public virtual bool ContainsKey(string key)
-            {
-                return IsValidIndex(IndexOfKey(key));
-            }
-
-            bool IList.Contains(object link)
-            {
-                if (link is Link)
-                {
-                    return Contains((Link)link);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            public int IndexOf(Link link)
-            {
-                return _owner._links.IndexOf(link);
-            }
-
-            int IList.IndexOf(object link)
-            {
-                if (link is Link)
-                {
-                    return IndexOf((Link)link);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-
-            /// <summary>
-            ///  The zero-based index of the first occurrence of value within the entire CollectionBase, if found; otherwise, -1.
-            /// </summary>
-            public virtual int IndexOfKey(string key)
-            {
-                // Step 0 - Arg validation
-                if (string.IsNullOrEmpty(key))
-                {
-                    return -1; // we dont support empty or null keys.
-                }
-
-                // step 1 - check the last cached item
-                if (IsValidIndex(lastAccessedIndex))
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[lastAccessedIndex].Name, key, /* ignoreCase = */ true))
-                    {
-                        return lastAccessedIndex;
-                    }
-                }
-
-                // step 2 - search for the item
-                for (int i = 0; i < Count; i++)
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[i].Name, key, /* ignoreCase = */ true))
-                    {
-                        lastAccessedIndex = i;
-                        return i;
-                    }
-                }
-
-                // step 3 - we didn't find it.  Invalidate the last accessed index and return -1.
-                lastAccessedIndex = -1;
-                return -1;
-            }
-
-            /// <summary>
-            ///  Determines if the index is valid for the collection.
-            /// </summary>
-            private bool IsValidIndex(int index)
-            {
-                return ((index >= 0) && (index < Count));
-            }
-
-            /// <summary>
-            ///  Remove all links from the linkLabel.
-            /// </summary>
-            public virtual void Clear()
-            {
-                bool doLayout = _owner._links.Count > 0 && _owner.AutoSize;
-                _owner._links.Clear();
-
-                if (doLayout)
-                {
-                    LayoutTransaction.DoLayout(_owner.ParentInternal, _owner, PropertyNames.Links);
-                    _owner.AdjustSize();
-                    _owner.Invalidate();
-                }
-
-                _owner.UpdateSelectability();
-                _owner.InvalidateTextLayout();
-                _owner.Invalidate();
-            }
-
-            void ICollection.CopyTo(Array dest, int index)
-            {
-                _owner._links.CopyTo(dest, index);
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                if (_owner._links != null)
-                {
-                    return _owner._links.GetEnumerator();
-                }
-                else
-                {
-                    return Array.Empty<Link>().GetEnumerator();
-                }
-            }
-
-            public void Remove(Link value)
-            {
-                if (value.Owner != _owner)
-                {
-                    return;
-                }
-
-                _owner._links.Remove(value);
-
-                if (_owner.AutoSize)
-                {
-                    LayoutTransaction.DoLayout(_owner.ParentInternal, _owner, PropertyNames.Links);
-                    _owner.AdjustSize();
-                    _owner.Invalidate();
-                }
-
-                _owner._links.Sort(LinkLabel.s_linkComparer);
-
-                _owner.ValidateNoOverlappingLinks();
-                _owner.UpdateSelectability();
-                _owner.InvalidateTextLayout();
-                _owner.Invalidate();
-
-                if (_owner.FocusLink is null && _owner._links.Count > 0)
-                {
-                    _owner.FocusLink = (Link)_owner._links[0];
-                }
-            }
-
-            public void RemoveAt(int index)
-            {
-                Remove(this[index]);
-            }
-
-            /// <summary>
-            ///  Removes the child control with the specified key.
-            /// </summary>
-            public virtual void RemoveByKey(string key)
-            {
-                int index = IndexOfKey(key);
-                if (IsValidIndex(index))
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            void IList.Remove(object value)
-            {
-                if (value is Link)
-                {
-                    Remove((Link)value);
-                }
-            }
-        }
-
-        [TypeConverter(typeof(LinkConverter))]
-        public class Link
-        {
-            private int _start;
-            private bool _enabled = true;
-            internal int _length;
-            private string _name;
-
-            public Link()
-            {
-            }
-
-            public Link(int start, int length)
-            {
-                _start = start;
-                _length = length;
-            }
-
-            public Link(int start, int length, object linkData)
-            {
-                _start = start;
-                _length = length;
-                LinkData = linkData;
-            }
-
-            internal Link(LinkLabel owner)
-            {
-                Owner = owner;
-            }
-
-            /// <summary>
-            ///  Description for accessibility
-            /// </summary>
-            public string Description { get; set; }
-
-            [DefaultValue(true)]
-            public bool Enabled
-            {
-                get => _enabled;
-                set
-                {
-                    if (_enabled != value)
-                    {
-                        _enabled = value;
-
-                        if ((int)(State & (LinkState.Hover | LinkState.Active)) != 0)
-                        {
-                            State &= ~(LinkState.Hover | LinkState.Active);
-                            if (Owner != null)
-                            {
-                                Owner.OverrideCursor = null;
-                            }
-                        }
-
-                        Owner?.InvalidateLink(this);
-                    }
-                }
-            }
-
-            public int Length
-            {
-                get
-                {
-                    if (_length == -1)
-                    {
-                        if (Owner != null && !string.IsNullOrEmpty(Owner.Text))
-                        {
-                            StringInfo stringInfo = new StringInfo(Owner.Text);
-                            return stringInfo.LengthInTextElements - Start;
-                        }
-                        else
-                        {
-                            return 0;
-                        }
-                    }
-
-                    return _length;
-                }
-                set
-                {
-                    if (_length != value)
-                    {
-                        _length = value;
-                        if (Owner != null)
-                        {
-                            Owner.InvalidateTextLayout();
-                            Owner.Invalidate();
-                        }
-                    }
-                }
-            }
-
-            [DefaultValue(null)]
-            public object LinkData { get; set; }
-
-            /// <summary>
-            ///  The LinkLabel object that owns this link.
-            /// </summary>
-            internal LinkLabel Owner { get; set; }
-
-            internal LinkState State { get; set; } = LinkState.Normal;
-
-            /// <summary>
-            ///  The name for the link - useful for indexing by key.
-            /// </summary>
-            [DefaultValue("")]
-            [SRCategory(nameof(SR.CatAppearance))]
-            [SRDescription(nameof(SR.TreeNodeNodeNameDescr))]
-            public string Name
-            {
-                get => _name ?? string.Empty;
-                set => _name = value;
-            }
-
-            public int Start
-            {
-                get => _start;
-                set
-                {
-                    if (_start != value)
-                    {
-                        _start = value;
-
-                        if (Owner != null)
-                        {
-                            Owner._links.Sort(LinkLabel.s_linkComparer);
-                            Owner.InvalidateTextLayout();
-                            Owner.Invalidate();
-                        }
-                    }
-                }
-            }
-
-            [SRCategory(nameof(SR.CatData))]
-            [Localizable(false)]
-            [Bindable(true)]
-            [SRDescription(nameof(SR.ControlTagDescr))]
-            [DefaultValue(null)]
-            [TypeConverter(typeof(StringConverter))]
-            public object Tag { get; set; }
-
-            [DefaultValue(false)]
-            public bool Visited
-            {
-                get => (State & LinkState.Visited) == LinkState.Visited;
-                set
-                {
-                    if (value != Visited)
-                    {
-                        if (value)
-                        {
-                            State |= LinkState.Visited;
-                        }
-                        else
-                        {
-                            State &= ~LinkState.Visited;
-                        }
-
-                        Owner?.InvalidateLink(this);
-                    }
-                }
-            }
-
-            internal Region VisualRegion { get; set; }
-        }
-
-        private class LinkComparer : IComparer
-        {
-            int IComparer.Compare(object link1, object link2)
-            {
-                Debug.Assert(link1 != null && link2 != null, "Null objects sent for comparison");
-
-                int pos1 = ((Link)link1).Start;
-                int pos2 = ((Link)link2).Start;
-
-                return pos1 - pos2;
-            }
-        }
-
-        internal class LinkLabelAccessibleObject : LabelAccessibleObject
-        {
-            public LinkLabelAccessibleObject(LinkLabel owner) : base(owner)
-            {
-            }
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            public override AccessibleObject GetChild(int index)
-            {
-                if (index >= 0 && index < ((LinkLabel)Owner).Links.Count)
-                {
-                    return new LinkAccessibleObject(((LinkLabel)Owner).Links[index]);
-                }
-                else
-                {
-                    return null;
-                }
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                if (propertyID == UiaCore.UIA.IsEnabledPropertyId)
-                {
-                    if (!Owner.Enabled)
-                    {
-                        return false;
-                    }
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
-
-            public override AccessibleObject HitTest(int x, int y)
-            {
-                if (!Owner.IsHandleCreated)
-                {
-                    return null;
-                }
-
-                Point p = Owner.PointToClient(new Point(x, y));
-                Link hit = ((LinkLabel)Owner).PointInLink(p.X, p.Y);
-
-                if (hit != null)
-                {
-                    return new LinkAccessibleObject(hit);
-                }
-
-                if (Bounds.Contains(x, y))
-                {
-                    return this;
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            /// </summary>
-            public override int GetChildCount()
-            {
-                return ((LinkLabel)Owner).Links.Count;
-            }
-        }
-
-        internal class LinkAccessibleObject : AccessibleObject
-        {
-            private readonly Link _link;
-
-            public LinkAccessibleObject(Link link) : base()
-            {
-                _link = link;
-            }
-
-            public override Rectangle Bounds
-            {
-                get
-                {
-                    if (!_link.Owner.IsHandleCreated)
-                    {
-                        return Rectangle.Empty;
-                    }
-
-                    Region region = _link.VisualRegion;
-                    Graphics g = Graphics.FromHwnd(_link.Owner.Handle);
-
-                    // Make sure we have a region for this link
-                    //
-                    if (region is null)
-                    {
-                        _link.Owner.EnsureRun(g);
-                        region = _link.VisualRegion;
-                        if (region is null)
-                        {
-                            g.Dispose();
-                            return Rectangle.Empty;
-                        }
-                    }
-
-                    Rectangle rect;
-                    try
-                    {
-                        rect = Rectangle.Ceiling(region.GetBounds(g));
-                    }
-                    finally
-                    {
-                        g.Dispose();
-                    }
-
-                    // Translate rect to screen coordinates
-                    //
-                    return _link.Owner.RectangleToScreen(rect);
-                }
-            }
-
-            public override string DefaultAction
-            {
-                get
-                {
-                    return SR.AccessibleActionClick;
-                }
-            }
-
-            public override string Description
-            {
-                get
-                {
-                    return _link.Description;
-                }
-            }
-
-            public override string Name
-            {
-                get
-                {
-                    string text = _link.Owner.Text;
-                    string name;
-
-                    // return the full name of the link label
-                    // as sometimes the link name in isolation
-                    // is unusable when using a screen reader
-                    name = text;
-                    if (_link.Owner.UseMnemonic)
-                    {
-                        name = WindowsFormsUtils.TextWithoutMnemonics(name);
-                    }
-
-                    return name;
-                }
-                set => base.Name = value;
-            }
-
-            public override AccessibleObject Parent
-            {
-                get
-                {
-                    return _link.Owner.AccessibilityObject;
-                }
-            }
-
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    return AccessibleRole.Link;
-                }
-            }
-
-            public override AccessibleStates State
-            {
-                get
-                {
-                    AccessibleStates state = AccessibleStates.Focusable;
-
-                    // Selected state
-                    //
-                    if (_link.Owner.FocusLink == _link)
-                    {
-                        state |= AccessibleStates.Focused;
-                    }
-
-                    return state;
-                }
-            }
-
-            public override string Value
-            {
-                get
-                {
-                    // Narrator announces Link's text twice, once as a Name property and once as a Value, thus removing value.
-                    // Value is optional for this role (Link).
-                    return string.Empty;
-                }
-            }
-
-            public override void DoDefaultAction()
-            {
-                _link.Owner.OnLinkClicked(new LinkLabelLinkClickedEventArgs(_link));
-            }
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                if (propertyID == UiaCore.UIA.IsEnabledPropertyId)
-                {
-                    if (!_link.Owner.Enabled)
-                    {
-                        return false;
-                    }
-                }
-
-                return base.GetPropertyValue(propertyID);
             }
         }
     }
