+++ src/System.Windows.Forms/src/System/Windows/Forms/ListView.cs
@@ -6944,2947 +6944,6 @@ namespace System.Windows.Forms
             };
         }
 
-        ///new class for comparing and sorting Icons ....
-        //subhag
-        internal class IconComparer : IComparer
-        {
-            private SortOrder sortOrder;
-
-            public IconComparer(SortOrder currentSortOrder)
-            {
-                sortOrder = currentSortOrder;
-            }
-
-            public SortOrder SortOrder
-            {
-                set
-                {
-                    sortOrder = value;
-                }
-            }
-
-            public int Compare(object obj1, object obj2)
-            {
-                //subhag
-                ListViewItem currentItem = (ListViewItem)obj1;
-                ListViewItem nextItem = (ListViewItem)obj2;
-                if (sortOrder == SortOrder.Ascending)
-                {
-                    return string.Compare(currentItem.Text, nextItem.Text, false, CultureInfo.CurrentCulture);
-                }
-                else
-                {
-                    return string.Compare(nextItem.Text, currentItem.Text, false, CultureInfo.CurrentCulture);
-                }
-            }
-        }
-        //end subhag
-
-        [ListBindable(false)]
-        public class CheckedIndexCollection : IList
-        {
-            private readonly ListView owner;
-
-            /* C#r: protected */
-            public CheckedIndexCollection(ListView owner)
-            {
-                this.owner = owner;
-            }
-
-            /// <summary>
-            ///  Number of currently selected items.
-            /// </summary>
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    if (!owner.CheckBoxes)
-                    {
-                        return 0;
-                    }
-
-                    // Count the number of checked items
-                    //
-                    int count = 0;
-                    foreach (ListViewItem item in owner.Items)
-                    {
-                        if (item != null && item.Checked)
-                        {
-                            count++;
-                        }
-                    }
-                    return count;
-                }
-            }
-
-            private int[] IndicesArray
-            {
-                get
-                {
-                    int[] indices = new int[Count];
-                    int index = 0;
-                    for (int i = 0; i < owner.Items.Count && index < indices.Length; ++i)
-                    {
-                        if (owner.Items[i].Checked)
-                        {
-                            indices[index++] = i;
-                        }
-                    }
-                    return indices;
-                }
-            }
-
-            /// <summary>
-            ///  Selected item in the list.
-            /// </summary>
-            public int this[int index]
-            {
-                get
-                {
-                    if (index < 0)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    // Loop through the main collection until we find the right index.
-                    //
-                    int cnt = owner.Items.Count;
-                    int nChecked = 0;
-                    for (int i = 0; i < cnt; i++)
-                    {
-                        ListViewItem item = owner.Items[i];
-
-                        if (item.Checked)
-                        {
-                            if (nChecked == index)
-                            {
-                                return i;
-                            }
-                            nChecked++;
-                        }
-                    }
-
-                    // Should never get to this point.
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    return this[index];
-                }
-                set
-                {
-                    throw new NotSupportedException();
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            public bool Contains(int checkedIndex)
-            {
-                if (owner.Items[checkedIndex].Checked)
-                {
-                    return true;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            bool IList.Contains(object checkedIndex)
-            {
-                if (checkedIndex is int)
-                {
-                    return Contains((int)checkedIndex);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            public int IndexOf(int checkedIndex)
-            {
-                int[] indices = IndicesArray;
-                for (int index = 0; index < indices.Length; ++index)
-                {
-                    if (indices[index] == checkedIndex)
-                    {
-                        return index;
-                    }
-                }
-                return -1;
-            }
-
-            int IList.IndexOf(object checkedIndex)
-            {
-                if (checkedIndex is int)
-                {
-                    return IndexOf((int)checkedIndex);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-
-            int IList.Add(object value)
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.Clear()
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.Remove(object value)
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.RemoveAt(int index)
-            {
-                throw new NotSupportedException();
-            }
-
-            void ICollection.CopyTo(Array dest, int index)
-            {
-                if (Count > 0)
-                {
-                    System.Array.Copy(IndicesArray, 0, dest, index, Count);
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                int[] indices = IndicesArray;
-                if (indices != null)
-                {
-                    return indices.GetEnumerator();
-                }
-                else
-                {
-                    return Array.Empty<int>().GetEnumerator();
-                }
-            }
-        }
-
-        [ListBindable(false)]
-        public class CheckedListViewItemCollection : IList
-        {
-            private readonly ListView owner;
-
-            ///  A caching mechanism for key accessor
-            ///  We use an index here rather than control so that we don't have lifetime
-            ///  issues by holding on to extra references.
-            private int lastAccessedIndex = -1;
-
-            /* C#r: protected */
-            public CheckedListViewItemCollection(ListView owner)
-            {
-                this.owner = owner;
-            }
-
-            /// <summary>
-            ///  Number of currently selected items.
-            /// </summary>
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                    }
-
-                    return owner.CheckedIndices.Count;
-                }
-            }
-
-            private ListViewItem[] ItemArray
-            {
-                get
-                {
-                    ListViewItem[] items = new ListViewItem[Count];
-                    int index = 0;
-                    for (int i = 0; i < owner.Items.Count && index < items.Length; ++i)
-                    {
-                        if (owner.Items[i].Checked)
-                        {
-                            items[index++] = owner.Items[i];
-                        }
-                    }
-                    return items;
-                }
-            }
-
-            /// <summary>
-            ///  Selected item in the list.
-            /// </summary>
-            public ListViewItem this[int index]
-            {
-                get
-                {
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                    }
-
-                    int itemIndex = owner.CheckedIndices[index];
-                    return owner.Items[itemIndex];
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                    }
-
-                    return this[index];
-                }
-                set
-                {
-                    throw new NotSupportedException();
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the child control with the specified key.
-            /// </summary>
-            public virtual ListViewItem this[string key]
-            {
-                get
-                {
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                    }
-
-                    // We do not support null and empty string as valid keys.
-                    if (string.IsNullOrEmpty(key))
-                    {
-                        return null;
-                    }
-
-                    // Search for the key in our collection
-                    int index = IndexOfKey(key);
-                    if (IsValidIndex(index))
-                    {
-                        return this[index];
-                    }
-                    else
-                    {
-                        return null;
-                    }
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            public bool Contains(ListViewItem item)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                }
-
-                if (item != null && item.ListView == owner && item.Checked)
-                {
-                    return true;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            bool IList.Contains(object item)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                }
-
-                if (item is ListViewItem)
-                {
-                    return Contains((ListViewItem)item);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Returns true if the collection contains an item with the specified key, false otherwise.
-            /// </summary>
-            public virtual bool ContainsKey(string key)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                }
-
-                return IsValidIndex(IndexOfKey(key));
-            }
-
-            public int IndexOf(ListViewItem item)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                }
-
-                ListViewItem[] items = ItemArray;
-                for (int index = 0; index < items.Length; ++index)
-                {
-                    if (items[index] == item)
-                    {
-                        return index;
-                    }
-                }
-                return -1;
-            }
-
-            /// <summary>
-            ///  The zero-based index of the first occurrence of value within the entire CollectionBase, if found; otherwise, -1.
-            /// </summary>
-            public virtual int IndexOfKey(string key)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                }
-
-                // Step 0 - Arg validation
-                if (string.IsNullOrEmpty(key))
-                {
-                    return -1; // we dont support empty or null keys.
-                }
-
-                // step 1 - check the last cached item
-                if (IsValidIndex(lastAccessedIndex))
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[lastAccessedIndex].Name, key, /* ignoreCase = */ true))
-                    {
-                        return lastAccessedIndex;
-                    }
-                }
-
-                // step 2 - search for the item
-                for (int i = 0; i < Count; i++)
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[i].Name, key, /* ignoreCase = */ true))
-                    {
-                        lastAccessedIndex = i;
-                        return i;
-                    }
-                }
-
-                // step 3 - we didn't find it.  Invalidate the last accessed index and return -1.
-                lastAccessedIndex = -1;
-                return -1;
-            }
-
-            /// <summary>
-            ///  Determines if the index is valid for the collection.
-            /// </summary>
-            private bool IsValidIndex(int index)
-            {
-                return (index >= 0) && (index < Count);
-            }
-
-            int IList.IndexOf(object item)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                }
-
-                if (item is ListViewItem)
-                {
-                    return IndexOf((ListViewItem)item);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-
-            int IList.Add(object value)
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.Clear()
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.Remove(object value)
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.RemoveAt(int index)
-            {
-                throw new NotSupportedException();
-            }
-
-            public void CopyTo(Array dest, int index)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                }
-
-                if (Count > 0)
-                {
-                    System.Array.Copy(ItemArray, 0, dest, index, Count);
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessCheckedItemsCollectionWhenInVirtualMode);
-                }
-
-                ListViewItem[] items = ItemArray;
-                if (items != null)
-                {
-                    return items.GetEnumerator();
-                }
-                else
-                {
-                    return Array.Empty<ListViewItem>().GetEnumerator();
-                }
-            }
-        }
-
-        [ListBindable(false)]
-        public class SelectedIndexCollection : IList
-        {
-            private readonly ListView owner;
-
-            /* C#r: protected */
-            public SelectedIndexCollection(ListView owner)
-            {
-                this.owner = owner;
-            }
-
-            /// <summary>
-            ///  Number of currently selected items.
-            /// </summary>
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    if (owner.IsHandleCreated)
-                    {
-                        return unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.GETSELECTEDCOUNT));
-                    }
-                    else
-                    {
-                        if (owner.savedSelectedItems != null)
-                        {
-                            return owner.savedSelectedItems.Count;
-                        }
-                        return 0;
-                    }
-                }
-            }
-
-            private int[] IndicesArray
-            {
-                get
-                {
-                    int count = Count;
-                    int[] indices = new int[count];
-
-                    if (owner.IsHandleCreated)
-                    {
-                        int displayIndex = -1;
-                        for (int i = 0; i < count; i++)
-                        {
-                            int fidx = unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.GETNEXTITEM, (IntPtr)displayIndex, (IntPtr)LVNI.SELECTED));
-                            if (fidx > -1)
-                            {
-                                indices[i] = fidx;
-                                displayIndex = fidx;
-                            }
-                            else
-                            {
-                                throw new InvalidOperationException(SR.SelectedNotEqualActual);
-                            }
-                        }
-                    }
-                    else
-                    {
-                        Debug.Assert(owner.savedSelectedItems != null || count == 0, "if the count of selectedItems is greater than 0 then the selectedItems should have been saved by now");
-                        for (int i = 0; i < count; i++)
-                        {
-                            indices[i] = owner.savedSelectedItems[i].Index;
-                        }
-                    }
-
-                    return indices;
-                }
-            }
-
-            /// <summary>
-            ///  Selected item in the list.
-            /// </summary>
-            public int this[int index]
-            {
-                get
-                {
-                    if (index < 0 || index >= Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    if (owner.IsHandleCreated)
-                    {
-                        // Count through the selected items in the ListView, until
-                        // we reach the 'index'th selected item.
-                        //
-                        int fidx = -1;
-                        for (int count = 0; count <= index; count++)
-                        {
-                            fidx = unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.GETNEXTITEM, (IntPtr)fidx, (IntPtr)LVNI.SELECTED));
-                            Debug.Assert(fidx != -1, "Invalid index returned from LVM_GETNEXTITEM");
-                        }
-
-                        return fidx;
-                    }
-                    else
-                    {
-                        Debug.Assert(owner.savedSelectedItems != null, "Null selected items collection");
-                        return owner.savedSelectedItems[index].Index;
-                    }
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    return this[index];
-                }
-                set
-                {
-                    throw new NotSupportedException();
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public bool Contains(int selectedIndex)
-            {
-                if (selectedIndex < 0 || selectedIndex >= owner.Items.Count)
-                {
-                    return false;
-                }
-
-                return owner.Items[selectedIndex].Selected;
-            }
-
-            bool IList.Contains(object selectedIndex)
-            {
-                if (selectedIndex is int)
-                {
-                    return Contains((int)selectedIndex);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            public int IndexOf(int selectedIndex)
-            {
-                int[] indices = IndicesArray;
-                for (int index = 0; index < indices.Length; ++index)
-                {
-                    if (indices[index] == selectedIndex)
-                    {
-                        return index;
-                    }
-                }
-                return -1;
-            }
-
-            int IList.IndexOf(object selectedIndex)
-            {
-                if (selectedIndex is int)
-                {
-                    return IndexOf((int)selectedIndex);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-
-            int IList.Add(object value)
-            {
-                if (value is int)
-                {
-                    return Add((int)value);
-                }
-                else
-                {
-                    throw new ArgumentException(string.Format(SR.InvalidArgument, nameof(value), value), nameof(value));
-                }
-            }
-
-            void IList.Clear()
-            {
-                Clear();
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                throw new NotSupportedException();
-            }
-
-            void IList.Remove(object value)
-            {
-                if (value is int)
-                {
-                    Remove((int)value);
-                }
-                else
-                {
-                    throw new ArgumentException(string.Format(SR.InvalidArgument, nameof(value), value), nameof(value));
-                }
-            }
-
-            void IList.RemoveAt(int index)
-            {
-                throw new NotSupportedException();
-            }
-
-            public int Add(int itemIndex)
-            {
-                if (owner.VirtualMode)
-                {
-                    if (itemIndex < 0 || itemIndex >= owner.VirtualListSize)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(itemIndex), itemIndex, string.Format(SR.InvalidArgument, nameof(itemIndex), itemIndex));
-                    }
-                    if (owner.IsHandleCreated)
-                    {
-                        owner.SetItemState(itemIndex, LVIS.SELECTED, LVIS.SELECTED);
-                        return Count;
-                    }
-                    else
-                    {
-                        return -1;
-                    }
-                }
-                else
-                {
-                    if (itemIndex < 0 || itemIndex >= owner.Items.Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(itemIndex), itemIndex, string.Format(SR.InvalidArgument, nameof(itemIndex), itemIndex));
-                    }
-                    owner.Items[itemIndex].Selected = true;
-                    return Count;
-                }
-            }
-
-            public void Clear()
-            {
-                if (!owner.VirtualMode)
-                {
-                    owner.savedSelectedItems = null;
-                }
-                if (owner.IsHandleCreated)
-                {
-                    owner.SetItemState(-1, 0, LVIS.SELECTED);
-                }
-            }
-
-            public void CopyTo(Array dest, int index)
-            {
-                if (Count > 0)
-                {
-                    System.Array.Copy(IndicesArray, 0, dest, index, Count);
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                int[] indices = IndicesArray;
-                if (indices != null)
-                {
-                    return indices.GetEnumerator();
-                }
-                else
-                {
-                    return Array.Empty<int>().GetEnumerator();
-                }
-            }
-
-            public void Remove(int itemIndex)
-            {
-                if (owner.VirtualMode)
-                {
-                    if (itemIndex < 0 || itemIndex >= owner.VirtualListSize)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(itemIndex), itemIndex, string.Format(SR.InvalidArgument, nameof(itemIndex), itemIndex));
-                    }
-                    if (owner.IsHandleCreated)
-                    {
-                        owner.SetItemState(itemIndex, 0, LVIS.SELECTED);
-                    }
-                }
-                else
-                {
-                    if (itemIndex < 0 || itemIndex >= owner.Items.Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(itemIndex), itemIndex, string.Format(SR.InvalidArgument, nameof(itemIndex), itemIndex));
-                    }
-                    owner.Items[itemIndex].Selected = false;
-                }
-            }
-        }
-
-        [ListBindable(false)]
-        public class SelectedListViewItemCollection : IList
-        {
-            private readonly ListView owner;
-
-            ///  A caching mechanism for key accessor
-            ///  We use an index here rather than control so that we don't have lifetime
-            ///  issues by holding on to extra references.
-            private int lastAccessedIndex = -1;
-
-            /* C#r: protected */
-            public SelectedListViewItemCollection(ListView owner)
-            {
-                this.owner = owner;
-            }
-
-            private ListViewItem[] SelectedItemArray
-            {
-                get
-                {
-                    if (owner.IsHandleCreated)
-                    {
-                        int cnt = unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.GETSELECTEDCOUNT));
-
-                        ListViewItem[] lvitems = new ListViewItem[cnt];
-
-                        int displayIndex = -1;
-
-                        for (int i = 0; i < cnt; i++)
-                        {
-                            int fidx = unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.GETNEXTITEM, (IntPtr)displayIndex, (IntPtr)LVNI.SELECTED));
-                            if (fidx > -1)
-                            {
-                                lvitems[i] = owner.Items[fidx];
-                                displayIndex = fidx;
-                            }
-                            else
-                            {
-                                throw new InvalidOperationException(SR.SelectedNotEqualActual);
-                            }
-                        }
-
-                        return lvitems;
-                    }
-                    else
-                    {
-                        if (owner.savedSelectedItems != null)
-                        {
-                            ListViewItem[] cloned = new ListViewItem[owner.savedSelectedItems.Count];
-                            for (int i = 0; i < owner.savedSelectedItems.Count; i++)
-                            {
-                                cloned[i] = owner.savedSelectedItems[i];
-                            }
-                            return cloned;
-                        }
-                        else
-                        {
-                            return Array.Empty<ListViewItem>();
-                        }
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Number of currently selected items.
-            /// </summary>
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                    }
-
-                    if (owner.IsHandleCreated)
-                    {
-                        return unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.GETSELECTEDCOUNT));
-                    }
-                    else
-                    {
-                        if (owner.savedSelectedItems != null)
-                        {
-                            return owner.savedSelectedItems.Count;
-                        }
-                        return 0;
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Selected item in the list.
-            /// </summary>
-            public ListViewItem this[int index]
-            {
-                get
-                {
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                    }
-
-                    if (index < 0 || index >= Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    if (owner.IsHandleCreated)
-                    {
-                        // Count through the selected items in the ListView, until
-                        // we reach the 'index'th selected item.
-                        //
-                        int fidx = -1;
-                        for (int count = 0; count <= index; count++)
-                        {
-                            fidx = unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.GETNEXTITEM, (IntPtr)fidx, (IntPtr)LVNI.SELECTED));
-                            Debug.Assert(fidx != -1, "Invalid index returned from LVM_GETNEXTITEM");
-                        }
-
-                        return owner.Items[fidx];
-                    }
-                    else
-                    {
-                        Debug.Assert(owner.savedSelectedItems != null, "Null selected items collection");
-                        return owner.savedSelectedItems[index];
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the child control with the specified key.
-            /// </summary>
-            public virtual ListViewItem this[string key]
-            {
-                get
-                {
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                    }
-
-                    // We do not support null and empty string as valid keys.
-                    if (string.IsNullOrEmpty(key))
-                    {
-                        return null;
-                    }
-
-                    // Search for the key in our collection
-                    int index = IndexOfKey(key);
-                    if (IsValidIndex(index))
-                    {
-                        return this[index];
-                    }
-                    else
-                    {
-                        return null;
-                    }
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                    }
-
-                    return this[index];
-                }
-                set
-                {
-                    // SelectedListViewItemCollection is read-only
-                    throw new NotSupportedException();
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            int IList.Add(object value)
-            {
-                // SelectedListViewItemCollection is read-only
-                throw new NotSupportedException();
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                // SelectedListViewItemCollection is read-only
-                throw new NotSupportedException();
-            }
-
-            /// <summary>
-            ///  Determines if the index is valid for the collection.
-            /// </summary>
-            private bool IsValidIndex(int index)
-            {
-                return (index >= 0) && (index < Count);
-            }
-
-            void IList.Remove(object value)
-            {
-                // SelectedListViewItemCollection is read-only
-                throw new NotSupportedException();
-            }
-
-            void IList.RemoveAt(int index)
-            {
-                // SelectedListViewItemCollection is read-only
-                throw new NotSupportedException();
-            }
-
-            /// <summary>
-            ///  Unselects all items.
-            /// </summary>
-            public void Clear()
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                ListViewItem[] items = SelectedItemArray;
-                for (int i = 0; i < items.Length; i++)
-                {
-                    items[i].Selected = false;
-                }
-            }
-
-            /// <summary>
-            ///  Returns true if the collection contains an item with the specified key, false otherwise.
-            /// </summary>
-            public virtual bool ContainsKey(string key)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                return IsValidIndex(IndexOfKey(key));
-            }
-
-            public bool Contains(ListViewItem item)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                return IndexOf(item) != -1;
-            }
-
-            bool IList.Contains(object item)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                if (item is ListViewItem)
-                {
-                    return Contains((ListViewItem)item);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            public void CopyTo(Array dest, int index)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                if (Count > 0)
-                {
-                    System.Array.Copy(SelectedItemArray, 0, dest, index, Count);
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                ListViewItem[] items = SelectedItemArray;
-                if (items != null)
-                {
-                    return items.GetEnumerator();
-                }
-                else
-                {
-                    return Array.Empty<ListViewItem>().GetEnumerator();
-                }
-            }
-
-            public int IndexOf(ListViewItem item)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                ListViewItem[] items = SelectedItemArray;
-                for (int index = 0; index < items.Length; ++index)
-                {
-                    if (items[index] == item)
-                    {
-                        return index;
-                    }
-                }
-                return -1;
-            }
-
-            int IList.IndexOf(object item)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                if (item is ListViewItem)
-                {
-                    return IndexOf((ListViewItem)item);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-
-            /// <summary>
-            ///  The zero-based index of the first occurrence of value within the entire CollectionBase, if found; otherwise, -1.
-            /// </summary>
-            public virtual int IndexOfKey(string key)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAccessSelectedItemsCollectionWhenInVirtualMode);
-                }
-
-                // Step 0 - Arg validation
-                if (string.IsNullOrEmpty(key))
-                {
-                    return -1; // we dont support empty or null keys.
-                }
-
-                // step 1 - check the last cached item
-                if (IsValidIndex(lastAccessedIndex))
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[lastAccessedIndex].Name, key, /* ignoreCase = */ true))
-                    {
-                        return lastAccessedIndex;
-                    }
-                }
-
-                // step 2 - search for the item
-                for (int i = 0; i < Count; i++)
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[i].Name, key, /* ignoreCase = */ true))
-                    {
-                        lastAccessedIndex = i;
-                        return i;
-                    }
-                }
-
-                // step 3 - we didn't find it.  Invalidate the last accessed index and return -1.
-                lastAccessedIndex = -1;
-                return -1;
-            }
-        }
-
-        [ListBindable(false)]
-        public class ColumnHeaderCollection : IList
-        {
-            private readonly ListView owner;
-
-            public ColumnHeaderCollection(ListView owner)
-            {
-                this.owner = owner ?? throw new ArgumentNullException(nameof(owner));
-            }
-
-            /// <summary>
-            ///  Given a Zero based index, returns the ColumnHeader object
-            ///  for the column at that index
-            /// </summary>
-            public virtual ColumnHeader this[int index]
-            {
-                get
-                {
-                    if (owner.columnHeaders is null || index < 0 || index >= owner.columnHeaders.Length)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    return owner.columnHeaders[index];
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    return this[index];
-                }
-                set
-                {
-                    throw new NotSupportedException();
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the child control with the specified key.
-            /// </summary>
-            public virtual ColumnHeader this[string key]
-            {
-                get
-                {
-                    // We do not support null and empty string as valid keys.
-                    if (string.IsNullOrEmpty(key))
-                    {
-                        return null;
-                    }
-
-                    // Search for the key in our collection
-                    int index = IndexOfKey(key);
-                    if (IsValidIndex(index))
-                    {
-                        return this[index];
-                    }
-                    else
-                    {
-                        return null;
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  The number of columns the ListView currently has in Details view.
-            /// </summary>
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    return owner.columnHeaders is null ? 0 : owner.columnHeaders.Length;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Removes the child control with the specified key.
-            /// </summary>
-            public virtual void RemoveByKey(string key)
-            {
-                int index = IndexOfKey(key);
-                if (IsValidIndex(index))
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            ///  A caching mechanism for key accessor
-            ///  We use an index here rather than control so that we don't have lifetime
-            ///  issues by holding on to extra references.
-            ///  Note this is not Thread Safe - but WinForms has to be run in a STA anyways.
-            private int lastAccessedIndex = -1;
-
-            /// <summary>
-            ///  The zero-based index of the first occurrence of value within the entire CollectionBase, if found; otherwise, -1.
-            /// </summary>
-            public virtual int IndexOfKey(string key)
-            {
-                // Step 0 - Arg validation
-                if (string.IsNullOrEmpty(key))
-                {
-                    return -1; // we dont support empty or null keys.
-                }
-
-                // step 1 - check the last cached item
-                if (IsValidIndex(lastAccessedIndex))
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[lastAccessedIndex].Name, key, /* ignoreCase = */ true))
-                    {
-                        return lastAccessedIndex;
-                    }
-                }
-
-                // step 2 - search for the item
-                for (int i = 0; i < Count; i++)
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[i].Name, key, /* ignoreCase = */ true))
-                    {
-                        lastAccessedIndex = i;
-                        return i;
-                    }
-                }
-
-                // step 3 - we didn't find it.  Invalidate the last accessed index and return -1.
-                lastAccessedIndex = -1;
-                return -1;
-            }
-
-            /// <summary>
-            ///  Determines if the index is valid for the collection.
-            /// </summary>
-            private bool IsValidIndex(int index)
-            {
-                return (index >= 0) && (index < Count);
-            }
-
-            /// <summary>
-            ///  Adds a column to the end of the Column list
-            /// </summary>
-            public virtual ColumnHeader Add(string text, int width, HorizontalAlignment textAlign)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Text = text,
-                    Width = width,
-                    TextAlign = textAlign
-                };
-                return owner.InsertColumn(Count, columnHeader);
-            }
-
-            public virtual int Add(ColumnHeader value)
-            {
-                int index = Count;
-                owner.InsertColumn(index, value);
-                return index;
-            }
-
-            public virtual ColumnHeader Add(string text)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Text = text
-                };
-                return owner.InsertColumn(Count, columnHeader);
-            }
-
-            // <-- NEW ADD OVERLOADS IN WHIDBEY
-
-            public virtual ColumnHeader Add(string text, int width)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Text = text,
-                    Width = width
-                };
-                return owner.InsertColumn(Count, columnHeader);
-            }
-
-            public virtual ColumnHeader Add(string key, string text)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Name = key,
-                    Text = text
-                };
-                return owner.InsertColumn(Count, columnHeader);
-            }
-
-            public virtual ColumnHeader Add(string key, string text, int width)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Name = key,
-                    Text = text,
-                    Width = width
-                };
-                return owner.InsertColumn(Count, columnHeader);
-            }
-
-            public virtual ColumnHeader Add(string key, string text, int width, HorizontalAlignment textAlign, string imageKey)
-            {
-                ColumnHeader columnHeader = new ColumnHeader(imageKey)
-                {
-                    Name = key,
-                    Text = text,
-                    Width = width,
-                    TextAlign = textAlign
-                };
-                return owner.InsertColumn(Count, columnHeader);
-            }
-
-            public virtual ColumnHeader Add(string key, string text, int width, HorizontalAlignment textAlign, int imageIndex)
-            {
-                ColumnHeader columnHeader = new ColumnHeader(imageIndex)
-                {
-                    Name = key,
-                    Text = text,
-                    Width = width,
-                    TextAlign = textAlign
-                };
-                return owner.InsertColumn(Count, columnHeader);
-            }
-
-            // END - NEW ADD OVERLOADS IN WHIDBEY  -->
-
-            public virtual void AddRange(ColumnHeader[] values)
-            {
-                if (values is null)
-                {
-                    throw new ArgumentNullException(nameof(values));
-                }
-
-                Hashtable usedIndices = new Hashtable();
-                int[] indices = new int[values.Length];
-
-                for (int i = 0; i < values.Length; i++)
-                {
-                    if (values[i] is null)
-                    {
-                        throw new ArgumentNullException(nameof(values));
-                    }
-
-                    if (values[i].DisplayIndex == -1)
-                    {
-                        values[i].DisplayIndexInternal = i;
-                    }
-
-                    if (!usedIndices.ContainsKey(values[i].DisplayIndex) && values[i].DisplayIndex >= 0 && values[i].DisplayIndex < values.Length)
-                    {
-                        usedIndices.Add(values[i].DisplayIndex, i);
-                    }
-
-                    indices[i] = values[i].DisplayIndex;
-                    Add(values[i]);
-                }
-
-                if (usedIndices.Count == values.Length)
-                {
-                    owner.SetDisplayIndices(indices);
-                }
-            }
-
-            int IList.Add(object value)
-            {
-                if (value is ColumnHeader)
-                {
-                    return Add((ColumnHeader)value);
-                }
-                else
-                {
-                    throw new ArgumentException(SR.ColumnHeaderCollectionInvalidArgument, nameof(value));
-                }
-            }
-
-            /// <summary>
-            ///  Removes all columns from the list view.
-            /// </summary>
-            public virtual void Clear()
-            {
-                // Delete the columns
-                if (owner.columnHeaders != null)
-                {
-                    if (owner.View == View.Tile)
-                    {
-                        // in Tile view our ListView uses the column header collection to update the Tile Information
-                        for (int colIdx = owner.columnHeaders.Length - 1; colIdx >= 0; colIdx--)
-                        {
-                            int w = owner.columnHeaders[colIdx].Width; // Update width before detaching from ListView
-                            owner.columnHeaders[colIdx].OwnerListview = null;
-                        }
-                        owner.columnHeaders = null;
-                        if (owner.IsHandleCreated)
-                        {
-                            owner.RecreateHandleInternal();
-                        }
-                    }
-                    else
-                    {
-                        for (int colIdx = owner.columnHeaders.Length - 1; colIdx >= 0; colIdx--)
-                        {
-                            int w = owner.columnHeaders[colIdx].Width; // Update width before detaching from ListView
-                            if (owner.IsHandleCreated)
-                            {
-                                User32.SendMessageW(owner, (User32.WM)LVM.DELETECOLUMN, (IntPtr)colIdx);
-                            }
-                            owner.columnHeaders[colIdx].OwnerListview = null;
-                        }
-                        owner.columnHeaders = null;
-                    }
-                }
-            }
-
-            public bool Contains(ColumnHeader value)
-            {
-                return IndexOf(value) != -1;
-            }
-
-            bool IList.Contains(object value)
-            {
-                if (value is ColumnHeader)
-                {
-                    return Contains((ColumnHeader)value);
-                }
-                return false;
-            }
-
-            /// <summary>
-            ///  Returns true if the collection contains an item with the specified key, false otherwise.
-            /// </summary>
-            public virtual bool ContainsKey(string key)
-            {
-                return IsValidIndex(IndexOfKey(key));
-            }
-
-            void ICollection.CopyTo(Array dest, int index)
-            {
-                if (Count > 0)
-                {
-                    System.Array.Copy(owner.columnHeaders, 0, dest, index, Count);
-                }
-            }
-
-            public int IndexOf(ColumnHeader value)
-            {
-                for (int index = 0; index < Count; ++index)
-                {
-                    if (this[index] == value)
-                    {
-                        return index;
-                    }
-                }
-                return -1;
-            }
-
-            int IList.IndexOf(object value)
-            {
-                if (value is ColumnHeader)
-                {
-                    return IndexOf((ColumnHeader)value);
-                }
-                return -1;
-            }
-
-            public void Insert(int index, ColumnHeader value)
-            {
-                if (index < 0 || index > Count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-                owner.InsertColumn(index, value);
-            }
-
-            void IList.Insert(int index, object value)
-            {
-                if (value is ColumnHeader)
-                {
-                    Insert(index, (ColumnHeader)value);
-                }
-            }
-
-            public void Insert(int index, string text, int width, HorizontalAlignment textAlign)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Text = text,
-                    Width = width,
-                    TextAlign = textAlign
-                };
-                Insert(index, columnHeader);
-            }
-
-            // <-- NEW INSERT OVERLOADS IN WHIDBEY
-
-            public void Insert(int index, string text)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Text = text
-                };
-                Insert(index, columnHeader);
-            }
-
-            public void Insert(int index, string text, int width)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Text = text,
-                    Width = width
-                };
-                Insert(index, columnHeader);
-            }
-
-            public void Insert(int index, string key, string text)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Name = key,
-                    Text = text
-                };
-                Insert(index, columnHeader);
-            }
-
-            public void Insert(int index, string key, string text, int width)
-            {
-                ColumnHeader columnHeader = new ColumnHeader
-                {
-                    Name = key,
-                    Text = text,
-                    Width = width
-                };
-                Insert(index, columnHeader);
-            }
-
-            public void Insert(int index, string key, string text, int width, HorizontalAlignment textAlign, string imageKey)
-            {
-                ColumnHeader columnHeader = new ColumnHeader(imageKey)
-                {
-                    Name = key,
-                    Text = text,
-                    Width = width,
-                    TextAlign = textAlign
-                };
-                Insert(index, columnHeader);
-            }
-
-            public void Insert(int index, string key, string text, int width, HorizontalAlignment textAlign, int imageIndex)
-            {
-                ColumnHeader columnHeader = new ColumnHeader(imageIndex)
-                {
-                    Name = key,
-                    Text = text,
-                    Width = width,
-                    TextAlign = textAlign
-                };
-                Insert(index, columnHeader);
-            }
-
-            // END - NEW INSERT OVERLOADS IN WHIDBEY -->
-
-            /// <summary>
-            ///  removes a column from the ListView
-            /// </summary>
-            public virtual void RemoveAt(int index)
-            {
-                if (owner.columnHeaders is null || index < 0 || index >= owner.columnHeaders.Length)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                int w = owner.columnHeaders[index].Width; // Update width before detaching from ListView
-
-                // in Tile view our ListView uses the column header collection to update the Tile Information
-                if (owner.IsHandleCreated && owner.View != View.Tile)
-                {
-                    int retval = unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.DELETECOLUMN, (IntPtr)index));
-                    if (0 == retval)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-                }
-
-                // we need to update the display indices
-                int[] indices = new int[Count - 1];
-
-                ColumnHeader removeHdr = this[index];
-                for (int i = 0; i < Count; i++)
-                {
-                    ColumnHeader hdr = this[i];
-                    if (i != index)
-                    {
-                        if (hdr.DisplayIndex >= removeHdr.DisplayIndex)
-                        {
-                            hdr.DisplayIndexInternal--;
-                        }
-                        indices[i > index ? i - 1 : i] = hdr.DisplayIndexInternal;
-                    }
-                }
-
-                removeHdr.DisplayIndexInternal = -1;
-
-                owner.columnHeaders[index].OwnerListview = null;
-                int columnCount = owner.columnHeaders.Length;
-                Debug.Assert(columnCount >= 1, "Column mismatch");
-                if (columnCount == 1)
-                {
-                    owner.columnHeaders = null;
-                }
-                else
-                {
-                    ColumnHeader[] newHeaders = new ColumnHeader[--columnCount];
-                    if (index > 0)
-                    {
-                        System.Array.Copy(owner.columnHeaders, 0, newHeaders, 0, index);
-                    }
-
-                    if (index < columnCount)
-                    {
-                        System.Array.Copy(owner.columnHeaders, index + 1, newHeaders, index, columnCount - index);
-                    }
-
-                    owner.columnHeaders = newHeaders;
-                }
-
-                // in Tile view our ListView uses the column header collection to update the Tile Information
-                if (owner.IsHandleCreated && owner.View == View.Tile)
-                {
-                    owner.RecreateHandleInternal();
-                }
-
-                owner.SetDisplayIndices(indices);
-            }
-
-            public virtual void Remove(ColumnHeader column)
-            {
-                int index = IndexOf(column);
-                if (index != -1)
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            void IList.Remove(object value)
-            {
-                if (value is ColumnHeader)
-                {
-                    Remove((ColumnHeader)value);
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                if (owner.columnHeaders != null)
-                {
-                    return owner.columnHeaders.GetEnumerator();
-                }
-                else
-                {
-                    return Array.Empty<ColumnHeader>().GetEnumerator();
-                }
-            }
-        }
-
-        /// <summary>
-        ///  Represents the collection of items in a ListView or ListViewGroup
-        /// </summary>
-        [ListBindable(false)]
-        public class ListViewItemCollection : IList
-        {
-            ///  A caching mechanism for key accessor
-            ///  We use an index here rather than control so that we don't have lifetime
-            ///  issues by holding on to extra references.
-            private int lastAccessedIndex = -1;
-
-            internal interface IInnerList
-            {
-                int Count { get; }
-                bool OwnerIsVirtualListView { get; }
-                bool OwnerIsDesignMode { get; }
-                ListViewItem this[int index] { get; set; }
-                ListViewItem Add(ListViewItem item);
-                void AddRange(ListViewItem[] items);
-                void Clear();
-                bool Contains(ListViewItem item);
-                void CopyTo(Array dest, int index);
-                IEnumerator GetEnumerator();
-                int IndexOf(ListViewItem item);
-                ListViewItem Insert(int index, ListViewItem item);
-                void Remove(ListViewItem item);
-                void RemoveAt(int index);
-            }
-
-            private readonly IInnerList innerList;
-
-            public ListViewItemCollection(ListView owner)
-            {
-                // Kept for APPCOMPAT reasons.
-                // In Whidbey this constructor is a no-op.
-
-                // initialize the inner list w/ a dummy list.
-                innerList = new ListViewNativeItemCollection(owner);
-            }
-
-            internal ListViewItemCollection(IInnerList innerList)
-            {
-                Debug.Assert(innerList != null, "Can't pass in null innerList");
-                this.innerList = innerList;
-            }
-
-            private IInnerList InnerList
-            {
-                get
-                {
-                    return innerList;
-                }
-            }
-
-            /// <summary>
-            ///  Returns the total number of items within the list view.
-            /// </summary>
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    return InnerList.Count;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Returns the ListViewItem at the given index.
-            /// </summary>
-            public virtual ListViewItem this[int index]
-            {
-                get
-                {
-                    if (index < 0 || index >= InnerList.Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    return InnerList[index];
-                }
-                set
-                {
-                    if (index < 0 || index >= InnerList.Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    InnerList[index] = value;
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    return this[index];
-                }
-                set
-                {
-                    if (value is ListViewItem)
-                    {
-                        this[index] = (ListViewItem)value;
-                    }
-                    else if (value != null)
-                    {
-                        this[index] = new ListViewItem(value.ToString(), -1);
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the child control with the specified key.
-            /// </summary>
-            public virtual ListViewItem this[string key]
-            {
-                get
-                {
-                    // We do not support null and empty string as valid keys.
-                    if (string.IsNullOrEmpty(key))
-                    {
-                        return null;
-                    }
-
-                    // Search for the key in our collection
-                    int index = IndexOfKey(key);
-                    if (IsValidIndex(index))
-                    {
-                        return this[index];
-                    }
-                    else
-                    {
-                        return null;
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Add an item to the ListView.  The item will be inserted either in
-            ///  the correct sorted position, or, if no sorting is set, at the end
-            ///  of the list.
-            /// </summary>
-            public virtual ListViewItem Add(string text)
-            {
-                return Add(text, -1);
-            }
-
-            int IList.Add(object item)
-            {
-                if (item is ListViewItem)
-                {
-                    return IndexOf(Add((ListViewItem)item));
-                }
-                else if (item != null)
-                {
-                    return IndexOf(Add(item.ToString()));
-                }
-                return -1;
-            }
-
-            /// <summary>
-            ///  Add an item to the ListView.  The item will be inserted either in
-            ///  the correct sorted position, or, if no sorting is set, at the end
-            ///  of the list.
-            /// </summary>
-            public virtual ListViewItem Add(string text, int imageIndex)
-            {
-                ListViewItem li = new ListViewItem(text, imageIndex);
-                Add(li);
-                return li;
-            }
-
-            /// <summary>
-            ///  Add an item to the ListView.  The item will be inserted either in
-            ///  the correct sorted position, or, if no sorting is set, at the end
-            ///  of the list.
-            /// </summary>
-            public virtual ListViewItem Add(ListViewItem value)
-            {
-                InnerList.Add(value);
-                return value;
-            }
-
-            // <-- NEW ADD OVERLOADS IN WHIDBEY
-
-            /// <summary>
-            ///  Add an item to the ListView.  The item will be inserted either in
-            ///  the correct sorted position, or, if no sorting is set, at the end
-            ///  of the list.
-            /// </summary>
-            public virtual ListViewItem Add(string text, string imageKey)
-            {
-                ListViewItem li = new ListViewItem(text, imageKey);
-                Add(li);
-                return li;
-            }
-
-            /// <summary>
-            ///  Add an item to the ListView.  The item will be inserted either in
-            ///  the correct sorted position, or, if no sorting is set, at the end
-            ///  of the list.
-            /// </summary>
-            public virtual ListViewItem Add(string key, string text, string imageKey)
-            {
-                ListViewItem li = new ListViewItem(text, imageKey)
-                {
-                    Name = key
-                };
-                Add(li);
-                return li;
-            }
-
-            /// <summary>
-            ///  Add an item to the ListView.  The item will be inserted either in
-            ///  the correct sorted position, or, if no sorting is set, at the end
-            ///  of the list.
-            /// </summary>
-            public virtual ListViewItem Add(string key, string text, int imageIndex)
-            {
-                ListViewItem li = new ListViewItem(text, imageIndex)
-                {
-                    Name = key
-                };
-                Add(li);
-                return li;
-            }
-
-            // END - NEW ADD OVERLOADS IN WHIDBEY  -->
-
-            public void AddRange(ListViewItem[] items)
-            {
-                if (items is null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-
-                InnerList.AddRange(items);
-            }
-
-            public void AddRange(ListViewItemCollection items)
-            {
-                if (items is null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-
-                ListViewItem[] itemArray = new ListViewItem[items.Count];
-                items.CopyTo(itemArray, 0);
-                InnerList.AddRange(itemArray);
-            }
-
-            /// <summary>
-            ///  Removes all items from the list view.
-            /// </summary>
-            public virtual void Clear()
-            {
-                InnerList.Clear();
-            }
-
-            public bool Contains(ListViewItem item)
-            {
-                return InnerList.Contains(item);
-            }
-
-            bool IList.Contains(object item)
-            {
-                if (item is ListViewItem)
-                {
-                    return Contains((ListViewItem)item);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Returns true if the collection contains an item with the specified key, false otherwise.
-            /// </summary>
-            public virtual bool ContainsKey(string key)
-            {
-                return IsValidIndex(IndexOfKey(key));
-            }
-
-            public void CopyTo(Array dest, int index)
-            {
-                InnerList.CopyTo(dest, index);
-            }
-
-            /// <summary>
-            ///  Searches for Controls by their Name property, builds up an array
-            ///  of all the controls that match.
-            /// </summary>
-            public ListViewItem[] Find(string key, bool searchAllSubItems)
-            {
-                if (string.IsNullOrEmpty(key))
-                {
-                    throw new ArgumentNullException(nameof(key), SR.FindKeyMayNotBeEmptyOrNull);
-                }
-
-                ArrayList foundItems = FindInternal(key, searchAllSubItems, this, new ArrayList());
-
-                ListViewItem[] stronglyTypedFoundItems = new ListViewItem[foundItems.Count];
-                foundItems.CopyTo(stronglyTypedFoundItems, 0);
-
-                return stronglyTypedFoundItems;
-            }
-
-            /// <summary>
-            ///  Searches for Controls by their Name property, builds up an arraylist
-            ///  of all the controls that match.
-            /// </summary>
-            private ArrayList FindInternal(string key, bool searchAllSubItems, ListViewItemCollection listViewItems, ArrayList foundItems)
-            {
-                if ((listViewItems is null) || (foundItems is null))
-                {
-                    return null;  //
-                }
-
-                for (int i = 0; i < listViewItems.Count; i++)
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(listViewItems[i].Name, key, /* ignoreCase = */ true))
-                    {
-                        foundItems.Add(listViewItems[i]);
-                    }
-                    else
-                    {
-                        if (searchAllSubItems)
-                        {
-                            // start from 1, as we've already compared subitems[0]
-                            for (int j = 1; j < listViewItems[i].SubItems.Count; j++)
-                            {
-                                if (WindowsFormsUtils.SafeCompareStrings(listViewItems[i].SubItems[j].Name, key, /* ignoreCase = */ true))
-                                {
-                                    foundItems.Add(listViewItems[i]);
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-
-                return foundItems;
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                if (InnerList.OwnerIsVirtualListView && !InnerList.OwnerIsDesignMode)
-                {
-                    // Throw the exception only at runtime.
-                    throw new InvalidOperationException(SR.ListViewCantGetEnumeratorInVirtualMode);
-                }
-                return InnerList.GetEnumerator();
-            }
-
-            public int IndexOf(ListViewItem item)
-            {
-                for (int index = 0; index < Count; ++index)
-                {
-                    if (this[index] == item)
-                    {
-                        return index;
-                    }
-                }
-                return -1;
-            }
-
-            int IList.IndexOf(object item)
-            {
-                if (item is ListViewItem)
-                {
-                    return IndexOf((ListViewItem)item);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-            /// <summary>
-            ///  The zero-based index of the first occurrence of value within the entire CollectionBase, if found; otherwise, -1.
-            /// </summary>
-            public virtual int IndexOfKey(string key)
-            {
-                // Step 0 - Arg validation
-                if (string.IsNullOrEmpty(key))
-                {
-                    return -1; // we dont support empty or null keys.
-                }
-
-                // step 1 - check the last cached item
-                if (IsValidIndex(lastAccessedIndex))
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[lastAccessedIndex].Name, key, /* ignoreCase = */ true))
-                    {
-                        return lastAccessedIndex;
-                    }
-                }
-
-                // step 2 - search for the item
-                for (int i = 0; i < Count; i++)
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[i].Name, key, /* ignoreCase = */ true))
-                    {
-                        lastAccessedIndex = i;
-                        return i;
-                    }
-                }
-
-                // step 3 - we didn't find it.  Invalidate the last accessed index and return -1.
-                lastAccessedIndex = -1;
-                return -1;
-            }
-
-            /// <summary>
-            ///  Determines if the index is valid for the collection.
-            /// </summary>
-            private bool IsValidIndex(int index)
-            {
-                return (index >= 0) && (index < Count);
-            }
-
-            public ListViewItem Insert(int index, ListViewItem item)
-            {
-                if (index < 0 || index > Count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-                InnerList.Insert(index, item);
-                return item;
-            }
-
-            public ListViewItem Insert(int index, string text)
-            {
-                return Insert(index, new ListViewItem(text));
-            }
-
-            public ListViewItem Insert(int index, string text, int imageIndex)
-            {
-                return Insert(index, new ListViewItem(text, imageIndex));
-            }
-
-            void IList.Insert(int index, object item)
-            {
-                if (item is ListViewItem)
-                {
-                    Insert(index, (ListViewItem)item);
-                }
-                else if (item != null)
-                {
-                    Insert(index, item.ToString());
-                }
-            }
-
-            // <-- NEW INSERT OVERLOADS IN WHIDBEY
-
-            public ListViewItem Insert(int index, string text, string imageKey)
-            {
-                return Insert(index, new ListViewItem(text, imageKey));
-            }
-
-            public virtual ListViewItem Insert(int index, string key, string text, string imageKey)
-            {
-                ListViewItem li = new ListViewItem(text, imageKey)
-                {
-                    Name = key
-                };
-                return Insert(index, li);
-            }
-
-            public virtual ListViewItem Insert(int index, string key, string text, int imageIndex)
-            {
-                ListViewItem li = new ListViewItem(text, imageIndex)
-                {
-                    Name = key
-                };
-                return Insert(index, li);
-            }
-
-            // END - NEW INSERT OVERLOADS IN WHIDBEY -->
-
-            /// <summary>
-            ///  Removes an item from the ListView
-            /// </summary>
-            public virtual void Remove(ListViewItem item)
-            {
-                InnerList.Remove(item);
-            }
-
-            /// <summary>
-            ///  Removes an item from the ListView
-            /// </summary>
-            public virtual void RemoveAt(int index)
-            {
-                if (index < 0 || index >= Count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                InnerList.RemoveAt(index);
-            }
-
-            /// <summary>
-            ///  Removes the child control with the specified key.
-            /// </summary>
-            public virtual void RemoveByKey(string key)
-            {
-                int index = IndexOfKey(key);
-                if (IsValidIndex(index))
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            void IList.Remove(object item)
-            {
-                if (item is null || !(item is ListViewItem))
-                {
-                    return;
-                }
-
-                Remove((ListViewItem)item);
-            }
-        }
-
-        // Overrides ListViewItemCollection methods and properties to automatically update
-        // the native listview.
-        //
-        internal class ListViewNativeItemCollection : ListViewItemCollection.IInnerList
-        {
-            private readonly ListView owner;
-
-            public ListViewNativeItemCollection(ListView owner)
-            {
-                this.owner = owner;
-            }
-
-            public int Count
-            {
-                get
-                {
-                    owner.ApplyUpdateCachedItems();
-                    if (owner.VirtualMode)
-                    {
-                        return owner.VirtualListSize;
-                    }
-                    else
-                    {
-                        return owner.itemCount;
-                    }
-                }
-            }
-
-            public bool OwnerIsVirtualListView
-            {
-                get
-                {
-                    return owner.VirtualMode;
-                }
-            }
-
-            public bool OwnerIsDesignMode
-            {
-                get
-                {
-                    return owner.DesignMode;
-                }
-            }
-
-            public ListViewItem this[int displayIndex]
-            {
-                get
-                {
-                    owner.ApplyUpdateCachedItems();
-
-                    if (owner.VirtualMode)
-                    {
-                        // if we are showing virtual items, we need to get the item from the user
-                        RetrieveVirtualItemEventArgs rVI = new RetrieveVirtualItemEventArgs(displayIndex);
-                        owner.OnRetrieveVirtualItem(rVI);
-                        rVI.Item.SetItemIndex(owner, displayIndex);
-                        return rVI.Item;
-                    }
-                    else
-                    {
-                        if (displayIndex < 0 || displayIndex >= owner.itemCount)
-                        {
-                            throw new ArgumentOutOfRangeException(nameof(displayIndex), displayIndex, string.Format(SR.InvalidArgument, nameof(displayIndex), displayIndex));
-                        }
-
-                        if (owner.IsHandleCreated && !owner.ListViewHandleDestroyed)
-                        {
-                            return (ListViewItem)owner.listItemsTable[DisplayIndexToID(displayIndex)];
-                        }
-                        else
-                        {
-                            Debug.Assert(owner.listItemsArray != null, "listItemsArray is null, but the handle isn't created");
-                            return (ListViewItem)owner.listItemsArray[displayIndex];
-                        }
-                    }
-                }
-                set
-                {
-                    owner.ApplyUpdateCachedItems();
-                    if (owner.VirtualMode)
-                    {
-                        throw new InvalidOperationException(SR.ListViewCantModifyTheItemCollInAVirtualListView);
-                    }
-
-                    if (displayIndex < 0 || displayIndex >= owner.itemCount)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(displayIndex), displayIndex, string.Format(SR.InvalidArgument, nameof(displayIndex), displayIndex));
-                    }
-
-                    if (owner.ExpectingMouseUp)
-                    {
-                        owner.ItemCollectionChangedInMouseDown = true;
-                    }
-
-                    RemoveAt(displayIndex);
-                    Insert(displayIndex, value);
-                }
-            }
-
-            public ListViewItem Add(ListViewItem value)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAddItemsToAVirtualListView);
-                }
-                else
-                {
-                    Debug.Assert(!owner.FlipViewToLargeIconAndSmallIcon || Count == 0, "the FlipView... bit is turned off after adding 1 item.");
-
-                    // PERF.
-                    // Get the Checked bit before adding it to the back end.
-                    // This saves a call into NativeListView to retrieve the real index.
-                    bool valueChecked = value.Checked;
-
-                    owner.InsertItems(owner.itemCount, new ListViewItem[] { value }, true);
-
-                    if (owner.IsHandleCreated && !owner.CheckBoxes && valueChecked)
-                    {
-                        owner.UpdateSavedCheckedItems(value, true /*addItem*/);
-                    }
-
-                    if (owner.ExpectingMouseUp)
-                    {
-                        owner.ItemCollectionChangedInMouseDown = true;
-                    }
-
-                    return value;
-                }
-            }
-
-            public void AddRange(ListViewItem[] values)
-            {
-                if (values is null)
-                {
-                    throw new ArgumentNullException(nameof(values));
-                }
-
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAddItemsToAVirtualListView);
-                }
-
-                IComparer comparer = owner.listItemSorter;
-                owner.listItemSorter = null;
-
-                Debug.Assert(!owner.FlipViewToLargeIconAndSmallIcon || Count == 0, "the FlipView... bit is turned off after adding 1 item.");
-
-                bool[] checkedValues = null;
-
-                if (owner.IsHandleCreated && !owner.CheckBoxes)
-                {
-                    // PERF.
-                    // Cache the Checked bit before adding the item to the list view.
-                    // This saves a bunch of calls to native list view when we want to UpdateSavedCheckedItems.
-                    //
-                    checkedValues = new bool[values.Length];
-                    for (int i = 0; i < values.Length; i++)
-                    {
-                        checkedValues[i] = values[i].Checked;
-                    }
-                }
-
-                try
-                {
-                    owner.BeginUpdate();
-                    owner.InsertItems(owner.itemCount, values, true);
-
-                    if (owner.IsHandleCreated && !owner.CheckBoxes)
-                    {
-                        for (int i = 0; i < values.Length; i++)
-                        {
-                            if (checkedValues[i])
-                            {
-                                owner.UpdateSavedCheckedItems(values[i], true /*addItem*/);
-                            }
-                        }
-                    }
-                }
-                finally
-                {
-                    owner.listItemSorter = comparer;
-                    owner.EndUpdate();
-                }
-
-                if (owner.ExpectingMouseUp)
-                {
-                    owner.ItemCollectionChangedInMouseDown = true;
-                }
-
-#pragma warning disable SA1408 // Conditional expressions should declare precedence
-                if (comparer != null || (owner.Sorting != SortOrder.None) && !owner.VirtualMode)
-#pragma warning restore SA1408 // Conditional expressions should declare precedence
-                {
-                    owner.Sort();
-                }
-            }
-
-            private int DisplayIndexToID(int displayIndex)
-            {
-                Debug.Assert(!owner.VirtualMode, "in virtual mode, this method does not make any sense");
-                if (owner.IsHandleCreated && !owner.ListViewHandleDestroyed)
-                {
-                    // Obtain internal index of the item
-                    var lvItem = new LVITEMW
-                    {
-                        mask = LVIF.PARAM,
-                        iItem = displayIndex
-                    };
-                    User32.SendMessageW(owner, (User32.WM)LVM.GETITEMW, IntPtr.Zero, ref lvItem);
-                    return (int)lvItem.lParam;
-                }
-                else
-                {
-                    return this[displayIndex].ID;
-                }
-            }
-
-            public void Clear()
-            {
-                if (owner.itemCount > 0)
-                {
-                    owner.ApplyUpdateCachedItems();
-
-                    if (owner.IsHandleCreated && !owner.ListViewHandleDestroyed)
-                    {
-                        // walk the items to see which ones are selected.
-                        // we use the LVM_GETNEXTITEM message to see what the next selected item is
-                        // so we can avoid checking selection for each one.
-                        //
-                        int count = owner.Items.Count;
-                        int nextSelected = (int)User32.SendMessageW(owner, (User32.WM)LVM.GETNEXTITEM, (IntPtr)(-1), (IntPtr)LVNI.SELECTED);
-                        for (int i = 0; i < count; i++)
-                        {
-                            ListViewItem item = owner.Items[i];
-                            Debug.Assert(item != null, "Failed to get item at index " + i.ToString(CultureInfo.InvariantCulture));
-                            if (item != null)
-                            {
-                                // if it's the one we're looking for, ask for the next one
-                                //
-                                if (i == nextSelected)
-                                {
-                                    item.StateSelected = true;
-                                    nextSelected = (int)User32.SendMessageW(owner, (User32.WM)LVM.GETNEXTITEM, (IntPtr)nextSelected, (IntPtr)LVNI.SELECTED);
-                                }
-                                else
-                                {
-                                    // otherwise it's false
-                                    //
-                                    item.StateSelected = false;
-                                }
-                                item.UnHost(i, false);
-                            }
-                        }
-                        Debug.Assert(owner.listItemsArray is null, "listItemsArray not null, even though handle created");
-
-                        User32.SendMessageW(owner, (User32.WM)LVM.DELETEALLITEMS);
-
-                        // There's a problem in the list view that if it's in small icon, it won't pick upo the small icon
-                        // sizes until it changes from large icon, so we flip it twice here...
-                        //
-                        if (owner.View == View.SmallIcon)
-                        {
-                            if (Application.ComCtlSupportsVisualStyles)
-                            {
-                                owner.FlipViewToLargeIconAndSmallIcon = true;
-                            }
-                            else
-                            {
-                                Debug.Assert(!owner.FlipViewToLargeIconAndSmallIcon, "we only set this when comctl 6.0 is loaded");
-                                owner.View = View.LargeIcon;
-                                owner.View = View.SmallIcon;
-                            }
-                        }
-                    }
-                    else
-                    {
-                        int count = owner.Items.Count;
-
-                        for (int i = 0; i < count; i++)
-                        {
-                            ListViewItem item = owner.Items[i];
-                            if (item != null)
-                            {
-                                item.UnHost(i, true);
-                            }
-                        }
-
-                        Debug.Assert(owner.listItemsArray != null, "listItemsArray is null, but the handle isn't created");
-                        owner.listItemsArray.Clear();
-                    }
-
-                    owner.listItemsTable.Clear();
-                    if (owner.IsHandleCreated && !owner.CheckBoxes)
-                    {
-                        owner.savedCheckedItems = null;
-                    }
-                    owner.itemCount = 0;
-
-                    if (owner.ExpectingMouseUp)
-                    {
-                        owner.ItemCollectionChangedInMouseDown = true;
-                    }
-                }
-            }
-
-            public bool Contains(ListViewItem item)
-            {
-                owner.ApplyUpdateCachedItems();
-                if (owner.IsHandleCreated && !owner.ListViewHandleDestroyed)
-                {
-                    return owner.listItemsTable[item.ID] == item;
-                }
-                else
-                {
-                    Debug.Assert(owner.listItemsArray != null, "listItemsArray is null, but the handle isn't created");
-                    return owner.listItemsArray.Contains(item);
-                }
-            }
-
-            public ListViewItem Insert(int index, ListViewItem item)
-            {
-                int count = 0;
-                if (owner.VirtualMode)
-                {
-                    count = Count;
-                }
-                else
-                {
-                    count = owner.itemCount;
-                }
-
-                if (index < 0 || index > count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantAddItemsToAVirtualListView);
-                }
-
-                Debug.Assert(!owner.FlipViewToLargeIconAndSmallIcon || Count == 0, "the FlipView... bit is turned off after adding 1 item.");
-
-                if (index < count)
-                {
-                    // if we're not inserting at the end, force the add.
-                    //
-                    owner.ApplyUpdateCachedItems();
-                }
-
-                owner.InsertItems(index, new ListViewItem[] { item }, true);
-                if (owner.IsHandleCreated && !owner.CheckBoxes && item.Checked)
-                {
-                    owner.UpdateSavedCheckedItems(item, true /*addItem*/);
-                }
-
-                if (owner.ExpectingMouseUp)
-                {
-                    owner.ItemCollectionChangedInMouseDown = true;
-                }
-
-                return item;
-            }
-
-            public int IndexOf(ListViewItem item)
-            {
-                Debug.Assert(!owner.VirtualMode, "in virtual mode, this function does not make any sense");
-                for (int i = 0; i < Count; i++)
-                {
-                    if (item == this[i])
-                    {
-                        return i;
-                    }
-                }
-                return -1;
-            }
-
-            public void Remove(ListViewItem item)
-            {
-                int index = owner.VirtualMode ? Count - 1 : IndexOf(item);
-
-                Debug.Assert(!owner.FlipViewToLargeIconAndSmallIcon || Count == 0, "the FlipView... bit is turned off after adding 1 item.");
-
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantRemoveItemsFromAVirtualListView);
-                }
-
-                if (index != -1)
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            public void RemoveAt(int index)
-            {
-                if (owner.VirtualMode)
-                {
-                    throw new InvalidOperationException(SR.ListViewCantRemoveItemsFromAVirtualListView);
-                }
-
-                if (index < 0 || index >= owner.itemCount)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                Debug.Assert(!owner.FlipViewToLargeIconAndSmallIcon || Count == 0, "the FlipView... bit is turned off after adding 1 item.");
-
-                if (owner.IsHandleCreated && !owner.CheckBoxes && this[index].Checked)
-                {
-                    owner.UpdateSavedCheckedItems(this[index], false /*addItem*/);
-                }
-
-                owner.ApplyUpdateCachedItems();
-                int itemID = DisplayIndexToID(index);
-                this[index].UnHost(true);
-
-                if (owner.IsHandleCreated)
-                {
-                    Debug.Assert(owner.listItemsArray is null, "listItemsArray not null, even though handle created");
-                    int retval = unchecked((int)(long)User32.SendMessageW(owner, (User32.WM)LVM.DELETEITEM, (IntPtr)index));
-
-                    if (0 == retval)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-                }
-                else
-                {
-                    Debug.Assert(owner.listItemsArray != null, "listItemsArray is null, but the handle isn't created");
-                    owner.listItemsArray.RemoveAt(index);
-                }
-
-                owner.itemCount--;
-                owner.listItemsTable.Remove(itemID);
-
-                if (owner.ExpectingMouseUp)
-                {
-                    owner.ItemCollectionChangedInMouseDown = true;
-                }
-            }
-
-            public void CopyTo(Array dest, int index)
-            {
-                if (owner.itemCount > 0)
-                {
-                    for (int displayIndex = 0; displayIndex < Count; ++displayIndex)
-                    {
-                        dest.SetValue(this[displayIndex], index++);
-                    }
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                ListViewItem[] items = new ListViewItem[owner.itemCount];
-                CopyTo(items, 0);
-
-                return items.GetEnumerator();
-            }
-        }
-
         /// <summary>
         ///  Creates the new instance of AccessibleObject for this ListView control.
         ///  Returning ListViewAccessibleObject.
